python/streamlit-langchain/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
memorystack>=1.0.0
python-dotenv>=1.0.0
